Refactor LeafletMap
* Refactor selected popup to ReconPortalSelectedPopUp and rework size issue in component to use update
* Refactor open topo datasets into their own component (OpenTopoLayer)
* Rename ReconPortalPopUp to ReconPortalPopUpContent
