+++ client/modules/reconportal/src/LeafletMap/LeafletMap.tsx
@@ -1,30 +1,26 @@
-import React, { useMemo, useState, useEffect } from 'react';
-import ReactDOM from 'react-dom/client';
+import React, { useMemo } from 'react';
 import {
   MapContainer,
   ZoomControl,
   TileLayer,
   LayersControl,
-  GeoJSON,
   Marker,
   Popup,
 } from 'react-leaflet';
-import { faMap, faLocationDot } from '@fortawesome/free-solid-svg-icons';
-import { OpenTopoPopup } from './OpenTopoPopup';
+import { faLocationDot } from '@fortawesome/free-solid-svg-icons';
 import {
   createSvgMarkerIcon,
   ZoomConditionalLayerGroup,
   createClusterIcon,
   ZoomOnEventSelection,
 } from './leafletUtil';
-import { getOpenTopoColor, getReconEventColor } from '../utils';
-import { getFirstLatLng } from './utils';
+import { getReconEventColor } from '../utils';
 import {
-  type ReconPortalEvent,
-  useGetOpenTopo,
   useReconEventContext,
   getReconPortalEventIdentifier,
+  ReconPortalEvent,
 } from '@client/hooks';
+import { OpenTopoLayer } from './OpenTopoLayer';
 import 'leaflet.markercluster';
 import MarkerClusterGroup from 'react-leaflet-markercluster';
 
@@ -33,8 +29,8 @@ import 'leaflet/dist/leaflet.css';
 */
 
 import styles from './LeafletMap.module.css';
-import { ReconPortalPopup } from './ReconPortalPopUp';
-import { LeafletMouseEvent } from 'leaflet';
+import { ReconPortalPopup } from './ReconPortalPopUpContent';
+import { ReconPortalSelectedPopup } from './ReconPortalSelectedPopUp';
 
 export const mapConfig = {
   startingCenter: [40, -80] as L.LatLngTuple,
@@ -51,159 +47,13 @@ export const mapConfig = {
  * Leaflet Map
  */
 export const LeafletMap: React.FC = () => {
-  const { data: openTopoData } = useGetOpenTopo();
   const {
     setSelectedReconPortalEventIdentifier,
     filteredReconPortalEvents,
     selectedReconPortalEventIdentifier,
   } = useReconEventContext();
 
-  const [showSelectedPopup, setShowSelectedPopup] = useState(false);
-
-  // Delay popup display to prevent incorrect initial sizing during zoom
-  useEffect(() => {
-    setShowSelectedPopup(false);
-
-    if (!selectedReconPortalEventIdentfier) return;
-
-    const timer = setTimeout(() => {
-      setShowSelectedPopup(true);
-    }, 1000);
-
-    return () => clearTimeout(timer);
-  }, [selectedReconPortalEventIdentfier]);
-
-  const openTopoMapFeatures = useMemo(() => {
-    const datasets = openTopoData?.Datasets ?? [];
-
-    // Features for open topo (just polygon and multipolygons)
-    const openTopoGeojsonFeatures: React.ReactNode[] = [];
-
-    // Markers for open topo things that aren't points
-    const openTopoMarkers: React.ReactNode[] = [];
-
-    // Fill openTopoGeojsonFeatures with everything
-    datasets.map(({ Dataset: dataset }, index) => {
-      const { geojson } = dataset.spatialCoverage.geo;
-
-      // Main GeoJSON rendering
-      openTopoGeojsonFeatures.push(
-        <GeoJSON
-          /* Duplicate datasets exist so need an extra index to make unqiue.  Background: OpenTopo seems to have \
-          a duplicate entry for some with multiple types of data e.g. "Point Clouds, Raster" */
-          key={`geojson-${dataset.identifier.value}-${index}`}
-          data={geojson}
-          style={() => ({
-            color: getOpenTopoColor(dataset),
-            weight: 2,
-            fillOpacity: 0.3,
-          })}
-          onEachFeature={(feature, layer) => {
-            const container = document.createElement('div');
-            layer.bindPopup(container);
-
-            let popupRendered = false;
-
-            const renderPopup = () => {
-              if (!popupRendered) {
-                ReactDOM.createRoot(container).render(
-                  <OpenTopoPopup dataset={dataset} />
-                );
-                popupRendered = true;
-
-                // Force Leaflet to recalculate popup size
-                // otherwise issue on first popup
-                setTimeout(() => {
-                  layer.getPopup()?.update();
-                }, 0);
-              }
-            };
-
-            const openPopupAtEventLocation = (e: LeafletMouseEvent) => {
-              renderPopup();
-              layer.openPopup(e.latlng);
-            };
-
-            // keep track of closing as when we move into popup that is a situation
-            // where we don't want to close it
-            let closeTimeout: ReturnType<typeof setTimeout> | null = null;
-
-            layer.on('mouseover', (e) => {
-              if (closeTimeout) {
-                clearTimeout(closeTimeout);
-                closeTimeout = null;
-              }
-              openPopupAtEventLocation(e);
-            });
-
-            layer.on('mouseout', () => {
-              closeTimeout = setTimeout(() => {
-                layer.closePopup();
-              }, 300);
-            });
-
-            /**
-             * Keeps the popup open when the user moves the mouse from the feature into the popup.
-             * Cancels the close timeout on mouseenter and restarts it on mouseleave.
-             * This prevents the popup from closing prematurely while the user interacts with links or content.
-             */
-            layer.on('popupopen', () => {
-              const popupEl = layer.getPopup()?.getElement();
-              if (!popupEl) return;
-
-              popupEl.addEventListener('mouseenter', () => {
-                if (closeTimeout) {
-                  clearTimeout(closeTimeout);
-                  closeTimeout = null;
-                }
-              });
-
-              popupEl.addEventListener('mouseleave', () => {
-                closeTimeout = setTimeout(() => {
-                  layer.closePopup();
-                }, 300);
-              });
-            });
-          }}
-        />
-      );
-    });
-
-    // Fill openTopoMarkers with a point for each dataset
-    datasets.forEach(({ Dataset: dataset }, index) => {
-      const { geojson } = dataset.spatialCoverage.geo;
-
-      const latlngPosition = getFirstLatLng(geojson);
-
-      const icon = createSvgMarkerIcon({
-        color: getOpenTopoColor(dataset),
-        icon: faMap,
-      });
-
-      openTopoMarkers.push(
-        <Marker
-          /* Duplicate datasets exist so need an extra index to make unqiue.  Background: OpenTopo seems to have \
-          a duplicate entry for some with multiple types of data e.g. "Point Clouds, Raster" */
-          key={`marker-${dataset.identifier.value}-${index}`}
-          position={latlngPosition}
-          icon={icon}
-          eventHandlers={{
-            mouseover: (e) => {
-              e.target.togglePopup();
-            },
-          }}
-        >
-          <Popup>
-            <OpenTopoPopup dataset={dataset} />
-          </Popup>
-        </Marker>
-      );
-    });
-
-    return [...openTopoGeojsonFeatures, ...openTopoMarkers];
-  }, [openTopoData]);
-
-  const handleFeatureClick = (reconEvent: ReconPortalEvents) => {
+  const handleFeatureClick = (reconEvent: ReconPortalEvent) => {
     setSelectedReconPortalEventIdentifier(
       getReconPortalEventIdentifier(reconEvent)
     );
@@ -221,6 +71,7 @@ export const LeafletMap: React.FC = () => {
   const ReconPortalEvents = useMemo(() => {
     const datasets = filteredReconPortalEvents ?? [];
     const reconPortalMarkers: React.ReactNode[] = [];
+    let closeTimeout: ReturnType<typeof setTimeout> | null = null;
 
     datasets.forEach((reconEvent, index) => {
       const icon = createSvgMarkerIcon({
@@ -233,13 +84,17 @@ export const LeafletMap: React.FC = () => {
           icon={icon}
           position={[reconEvent.location.lat, reconEvent.location.lon]}
           eventHandlers={{
-            click: (e) => handleFeatureClick(reconEvent),
+            click: () => handleFeatureClick(reconEvent),
             mouseover: (e) => {
               // Only show hover popup if this marker is not currently selected
               if (
                 getReconPortalEventIdentifier(reconEvent) !==
-                selectedReconPortalEventIdentfier
+                selectedReconPortalEventIdentifier
               ) {
+                if (closeTimeout) {
+                  clearTimeout(closeTimeout);
+                  closeTimeout = null;
+                }
                 e.target.openPopup();
               }
             },
@@ -247,12 +102,33 @@ export const LeafletMap: React.FC = () => {
               // Only close hover popup if this marker is not currently selected
               if (
                 getReconPortalEventIdentifier(reconEvent) !==
-                selectedReconPortalEventIdentfier
+                selectedReconPortalEventIdentifier
               ) {
-                setTimeout(() => {
+                closeTimeout = setTimeout(() => {
                   e.target.closePopup();
-                }, 1000);
+                }, 300);
               }
+            },
+
+            popupopen: (e) => {
+              // With these listeners, we are preventing the popup from closing if the
+              // user moves their mouse from the marker into the popup itself
+
+              const popupEl = e.popup.getElement();
+              if (!popupEl) return;
+
+              popupEl.addEventListener('mouseenter', () => {
+                if (closeTimeout) {
+                  clearTimeout(closeTimeout);
+                  closeTimeout = null;
+                }
+              });
+
+              popupEl.addEventListener('mouseleave', () => {
+                closeTimeout = setTimeout(() => {
+                  e.popup.close();
+                }, 300);
+              });
             },
           }}
         >
@@ -299,7 +175,7 @@ export const LeafletMap: React.FC = () => {
 
         {/* Open Topo features, only visible when zoomed in and DS event selected*/}
         <ZoomConditionalLayerGroup minZoom={9}>
-          {openTopoMapFeatures}
+          <OpenTopoLayer />
         </ZoomConditionalLayerGroup>
         {/* Recon Portal features, only zoomed in when DS event selected*/}
         <ZoomOnEventSelection zoomLevel={11}></ZoomOnEventSelection>
@@ -322,17 +198,8 @@ export const LeafletMap: React.FC = () => {
         <ZoomControl position="topright" />
 
         {/* Selected event banner popup */}
-        {selectedEvent && showSelectedPopup && (
-          <Popup
-            className={styles.selectedEventPopup}
-            position={[selectedEvent.location.lat, selectedEvent.location.lon]}
-            offset={[0, -10]}
-            closeButton={false}
-            closeOnClick={false}
-            autoClose={false}
-          >
-            <ReconPortalPopup dataset={selectedEvent} showDetails={false} />
-          </Popup>
+        {selectedEvent && (
+          <ReconPortalSelectedPopup selectedEvent={selectedEvent} />
         )}
       </MapContainer>
     </>
